move filter components into own file
